+++ test/feedback_test.dart
@@ -10,7 +10,7 @@ import 'test_app.dart';
 
 void main() {
   group('BetterFeedback', () {
-    testWidgets(' can open feedback', (tester) async {
+    testWidgets(' can open feedback with default settings', (tester) async {
       final widget = BetterFeedback(
         child: Builder(
           builder: (context) {
@@ -33,8 +33,40 @@ void main() {
       await tester.pumpAndSettle();
 
       userInputFields = find.byKey(const Key('feedback_bottom_sheet'));
+      final activeDrawingColor = getActiveColorButton();
 
       expect(userInputFields, findsOneWidget);
+      expect(activeDrawingColor, findsNothing);
+    });
+
+    testWidgets(' can open feedback in drawing mode', (tester) async {
+      final widget = BetterFeedback(
+        defaultNavigate: false,
+        child: Builder(
+          builder: (context) {
+            return const MyTestApp();
+          },
+        ),
+      );
+
+      await tester.pumpWidget(widget);
+      await tester.pumpAndSettle();
+
+      // feedback is closed
+      var userInputFields = find.byKey(const Key('feedback_bottom_sheet'));
+
+      expect(userInputFields, findsNothing);
+
+      // open feedback
+      final openFeedbackButton = find.text('open feedback');
+      await tester.tap(openFeedbackButton);
+      await tester.pumpAndSettle();
+
+      userInputFields = find.byKey(const Key('feedback_bottom_sheet'));
+      final activeDrawingColor = getActiveColorButton();
+
+      expect(userInputFields, findsOneWidget);
+      expect(activeDrawingColor.evaluate().length, 4);
     });
 
     testWidgets(' can close feedback', (tester) async {
@@ -119,4 +151,15 @@ class MockScreenshotController implements ScreenshotController {
       Duration delay = const Duration(milliseconds: 20)}) {
     return Future.value(Uint8List.fromList([1, 1, 1, 1]));
   }
+}
+
+Finder getActiveColorButton() {
+  return find.byWidgetPredicate((widget) {
+    if (widget is IconButton) {
+      final IconButton selectButton = widget;
+      return selectButton.color != null && selectButton.onPressed != null;
+    } else {
+      return false;
+    }
+  });
 }