+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.1.0]
+
+- Fix locale for Flutter Web (#13, #50)
+- Add option for changing the default navigation mode
+
 ## [1.0.2]
 
 - Fix: Color Icon buttons are bigger to improve accessibility